Dynamically load DB configs
