+++ src/config/databases.json
@@ -1,0 +1,1 @@
+["bindle"]
